backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "bcrypt==3.2.0",
    "certifi==2025.8.3",
    "cffi==1.17.1",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "cryptography==45.0.6",
    "ecdsa==0.19.1",
    "fastapi==0.116.1",
    "feedparser==6.0.12",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "idna==3.10",
    "passlib[bcrypt]==1.7.4",
    "pillow==11.3.0",
    "pyasn1==0.6.1",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "python-crontab==3.3.0",
    "python-jose[cryptography]==3.5.0",
    "python-mpd2==3.1.1",
    "python-multipart==0.0.20",
    "requests==2.32.5",
    "rsa==4.9.1",
    "sgmllib3k==1.0.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.42",
    "starlette==0.47.2",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "yt-dlp>=2025.12.8",
]
